refactor: reorganize renderers import paths and create index file for cleaner structure. Lazy load some on clientPage
